Move Flask app into api folder for Vercel
