+++ api/static/scripts/main.js
@@ -1,0 +1,1 @@
+// JavaScript placeholder
