+++ vercel.json
@@ -1,0 +1,8 @@
+{
+  "builds": [
+    {"src": "api/hello.py", "use": "@vercel/python"}
+  ],
+  "routes": [
+    {"src": "/(.*)", "dest": "api/hello.py"}
+  ]
+}
